+++ app/client/src/PluginActionEditor/hooks/index.ts
@@ -1,1 +1,3 @@
 export { useActionSettingsConfig } from "ee/PluginActionEditor/hooks/useActionSettingsConfig";
+export { useHandleDeleteClick } from "ee/PluginActionEditor/hooks/useHandleDeleteClick";
+export { useHandleRunClick } from "ee/PluginActionEditor/hooks/useHandleRunClick";
