--- app/client/src/ee/PluginActionEditor/hooks/useActionDispatchCalls.ts
@@ -1,1 +1,0 @@
-export * from "ce/PluginActionEditor/hooks/useActionDispatchCalls";
