+++ app/client/src/ee/PluginActionEditor/hooks/useHandleDeleteClick.ts
@@ -1,0 +1,1 @@
+export * from "ce/PluginActionEditor/hooks/useHandleDeleteClick";
